feat(analytics): Fix syntax for analytics with react-ga4
