+++ src/components/App/App.tsx
@@ -12,9 +12,17 @@ function App() {
   const isProduction = import.meta.env.PROD;
   const analyticsEnabled = useAppSelector(getAnalyticsEnabled);
   if (isProduction && analyticsEnabled) {
-    ReactGA.initialize("UA-1018242-68"); // TODO: should use the GA4 Measurement ID (which starts with "G-")
+    ReactGA.initialize("G-JHXHM8VXJ1", {
+      gtagOptions: {
+        custom_map: {
+          dimension1: "dashboardVersion",
+          dimension2: "controllerVersion",
+          dimension3: "isJuju",
+        },
+      },
+    });
     ReactGA.send({
-      hitType: "page_view",
+      hitType: "pageview",
       page: window.location.href.replace(window.location.origin, ""),
     });
   }
