+++ src/hooks/useAnalytics.test.ts
@@ -1,9 +1,9 @@
-import { renderHook } from "@testing-library/react";
 import ReactGA from "react-ga4";
 import type { MockInstance } from "vitest";
 import { vi } from "vitest";
 
 import * as store from "store/store";
+import { renderWrappedHook } from "testing/utils";
 
 import useAnalytics from "./useAnalytics";
 
@@ -38,7 +38,7 @@ describe("useAnalytics", () => {
       vi.fn().mockReturnValue(true),
     );
     vi.stubEnv("PROD", false);
-    const { result } = renderHook(() => useAnalytics());
+    const { result } = renderWrappedHook(() => useAnalytics());
     result.current({ path: "/some/path" });
     expect(eventSpy).not.toHaveBeenCalled();
     expect(pageviewSpy).not.toHaveBeenCalled();
@@ -48,7 +48,7 @@ describe("useAnalytics", () => {
     vi.spyOn(store, "useAppSelector").mockImplementation(
       vi.fn().mockReturnValue(false),
     );
-    const { result } = renderHook(() => useAnalytics());
+    const { result } = renderWrappedHook(() => useAnalytics());
     result.current({ path: "/some/path" });
     expect(eventSpy).not.toHaveBeenCalled();
     expect(pageviewSpy).not.toHaveBeenCalled();
@@ -58,10 +58,10 @@ describe("useAnalytics", () => {
     vi.spyOn(store, "useAppSelector").mockImplementation(
       vi.fn().mockReturnValue(true),
     );
-    const { result } = renderHook(() => useAnalytics());
+    const { result } = renderWrappedHook(() => useAnalytics());
     result.current({ path: "/some/path" });
     expect(pageviewSpy).toHaveBeenCalledWith({
-      hitType: "page_view",
+      hitType: "pageview",
       page: "/some/path",
     });
   });
@@ -70,11 +70,14 @@ describe("useAnalytics", () => {
     vi.spyOn(store, "useAppSelector").mockImplementation(
       vi.fn().mockReturnValue(true),
     );
-    const { result } = renderHook(() => useAnalytics());
+    const { result } = renderWrappedHook(() => useAnalytics());
     result.current({ category: "sidebar", action: "toggle" });
     expect(eventSpy).toHaveBeenCalledWith({
       category: "sidebar",
       action: "toggle",
+      controllerVersion: "",
+      dashboardVersion: "",
+      isJuju: "false",
     });
   });
 });
