+++ src/hooks/useAnalytics.tsx
@@ -1,34 +1,46 @@
 import ReactGA from "react-ga4";
+import { useSelector } from "react-redux";
 
-import { getAnalyticsEnabled } from "store/general/selectors";
+import {
+  getAnalyticsEnabled,
+  getAppVersion,
+  getControllerConnection,
+  getIsJuju,
+  getWSControllerURL,
+} from "store/general/selectors";
 import { useAppSelector } from "store/store";
 
 type AnalyticMessage = {
   path?: string;
   category?: string;
   action?: string;
-  eventParams?: { [key: string]: string };
 };
 
 export default function useAnalytics() {
   const analyticsEnabled = useAppSelector(getAnalyticsEnabled);
-  return ({
-    path,
-    category = "",
-    action = "",
-    eventParams = {},
-  }: AnalyticMessage) => {
+  const isJuju = useSelector(getIsJuju);
+  const appVersion = useSelector(getAppVersion);
+  const wsControllerURL = useAppSelector(getWSControllerURL);
+  const controllerVersion = useAppSelector((state) =>
+    getControllerConnection(state, wsControllerURL),
+  )?.serverVersion;
+
+  return ({ path, category = "", action = "" }: AnalyticMessage) => {
     const isProduction = import.meta.env.PROD;
     if (!isProduction || !analyticsEnabled) {
       return;
     }
     if (path) {
-      ReactGA.send({ hitType: "page_view", page: path });
+      ReactGA.send({ hitType: "pageview", page: path });
     } else {
       ReactGA.event({
         category,
         action,
-        ...eventParams,
+        ...{
+          dashboardVersion: appVersion ?? "",
+          controllerVersion: controllerVersion ?? "",
+          isJuju: (!!isJuju).toString(),
+        },
       });
     }
   };
